refactor(public): move getLoanOrders method
