+++ index.js
@@ -1,5 +1,4 @@
 module.exports = {
-  AuthenticatedClient: require('./lib/authenticated.js'),
-  PublicClient: require('./lib/public.js'),
-  WebsocketClient: require('./lib/websocket.js'),
+  AuthenticatedClient: require("./lib/authenticated.js"),
+  WebsocketClient: require("./lib/websocket.js")
 };
